--- src/utils/destinations.js
@@ -1,4 +1,0 @@
-export const destinations = () => ([
-  'Podgorica', 'Moscow', 'New York',
-  'Bratislava', 'Oslo', 'Ottawa', 'Prague'
-]);
