--- src/utils/waypointTypes.js
@@ -1,4 +1,0 @@
-export const wayPointTypes = () => ([
-  'Taxi', 'Bus', 'Train', 'Ship',
-  'Drive', 'Flight', 'Check-in', 'Sightseeing', 'Restaurant'
-]);
