+++ src/view/Abstract-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render';
+import { createElement } from '../utils/render.js';
 
 export default class AbstractView {
   #element = null;
